pyproject.toml
[project]
name = "llm-reviewer"
version = "0.1.0"
description = "LLM-based academic paper review pipeline"
requires-python = ">=3.12"
dependencies = [
    "coarse-ink>=1.3.0",
    "pymupdf4llm>=1.27.2.2",
    "python-dotenv>=1.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reviewer"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
